Add yamlgroup inventory plugin
The constructed inventory plugin allows expressing additional groups,
but it's too heavy weight for our needs. Additionally, it is a full
inventory plugin that will add hosts to the inventory if they don't
exist.

What we want instead is something that will associate existing hosts
(that would have come from another source) with groups.

Also, while we're at it, take our disabled concept and encode it
directly explicitly removing the disabled hosts from the inventory so we
don't need !disabled on every line.

This also switches to using emergency.yaml instead of emergency, which
uses the same format.

Change-Id: Iea8b2eb2e9c723aca06f75d3d3307893e320cced