Treat unparsed inventory as a failure
We don't want to run ansible if we don't get a complete inventory from
our clouds. The reason for this is we cannot be sure that the ordering
of git servers, gerrit, and zuul or our serialized updates of afs
servers will work correctly if we have an incomplete inventory.

Instead we just want ansible to fail and try again in the future (we can
then debug why our clouds are not working).

From the ansible docs for any_unparsed_is_failed:

  If 'true', it is a fatal error when any given inventory source
  cannot be successfully parsed by any available inventory plugin;
  otherwise, this situation only attracts a warning.

Additionally we tell openstack inventory plugin to report failures
rather than empty inventory so that the unparsed failures happen.

Change-Id: I9025776af4316fbdd2c910566883eb3a2530852a