Support any URI test result links in gerrit
Currently gerrit will markup an http[s] url into an anchor link. This is
then matched against when being reformatted into our comment_test
results list. This results list from the comments is then copied into
the results table at the top of the page.

When a uri such as a 'finger://' link is present in the CI results,
these are not formatted as anchors by gerrit. Thus these are missed as
being placed into the comment_test list and therefore the results table.

This change implements support for any URI in the CI results. It will
also convert it into an anchor link.

Change-Id: I62a2c0a944f3d51fdd9300008a8d0dff52c5e82f