Create the OpenStack discussion mailing list
In order to bring the operator/user/developer-focused segments of
our community together, we need a new discussion mailing list which
subsumes the topics of the current general, -dev, -operator and
-sigs mailing lists. This list will begin with configuration to
reject messages sent to it, but as we get closer to sunsetting the
original lists we'll subscribe it to them for a transitional period
and open it for discussion at that time.

I've set my E-mail address as the initial owner, but will coordinate
to find a team of list admins willing to take it on before it's
opened for posting.

Change-Id: I7dd62482229035c9b304573a3d93ef397cf43749