Manage user ssh keys from urls
Optionally allow for multiple ssh keys to be specified from urls. This
uses a small python ansible module to read all of the keys and construct
a single key to be used. This allows us to take advantage of the
authorized_key exclusive property while managing multiple keys from
urls.

Change-Id: Ib228056394964b1bfb4454f739c2395e8a8bf619