Initial port of install-docker role
This is an initial port of z-j:install-docker role to system-config
for use on the control-plane hosts.

There was some discussion over where to put this; the z-j role is
focused on being useful to CI jobs, and we don't wish to import z-j
into system-config production playbooks.  An external repo was
considered, but it was felt this was both fairly simple for now, and
has potential to end up fairly specific to OpenStack infra production
environments, making it unsuitable as a fully external role.

Change-Id: Ia4de5032789ff0f2b07d4f93c0c52cf94aa9c25c