Import install-docker role
This is a role for installing docker on our control-plane servers.

It is based on install-docker from zuul-jobs.

Basic testinfra tests are added; because docker fiddles the iptables
rules in magic ways, the firewall testing is moved out of the base
tests and skipped for now on the docker host.

Change-Id: Ia4de5032789ff0f2b07d4f93c0c52cf94aa9c25c