Retain fewer etherpad database backups
These backups were kept locally for convenience of operation. Bup stores
blobs in git so it should be effectively deduping these over time so
wasn't a stress on remote backups. However now we are filling the local
disk with the db backup files.

Address this by reducing the number of local backups to 5 (from 10).

Change-Id: I1765836784752dba2a6f2582801b860fcce4843c