Increase sync frequency for OpenSUSE repo
The OpenStack Ansible project relies on Infra's OpenSUSE repo mirror
for their OpenSUSE gate checks. It is therefore important that the
mirror is somewhat up to date and reflects a recent state of the
official repository.

In order to make "recent" more recent this change drops the sync
frequency for the OpenSUSE repositories to every 2 hours down from
every 6 hours.

Change-Id: I94012c7981cc8fd8c8da6e5fb4af3d8e402c9758