Cleanup Citycloud Lon1
This region does not show up in catalog listings anymore and is causing
inventory generation for ansible to fail. This change removes Lon1 from
the management side of things so that we can get ansible and puppet
running again.

This does not cleanup nodepool which we can do in a followup once
ansible and puppet are running again.

Change-Id: Icf3b19381ebba3498dfc204a48dc1ea52ae9d951