lists: run mailman verp router before dnslookup
The mailman verp router handles remote addresses like dnslookup.
It needs to run before dnslookup in order to be effective, so run
it first.  It's only for outgoing messages, not incoming, so won't
affect the blackhole aliases we have for incoming fake bounce
messages.

Note that the verp router hasn't been used in about a year due to
this oversight, so we should merge this change with caution.

Change-Id: I7d2a0f05f82485a54c1e7048f09b4edf6e0f0612