Don't install testinfra 1.17.0
testinfra 1.17.0 has a broken wheel on pypi that attempts to install
importlib regardless of the python version. Unfortunately this breaks
under python3 (and possibly under python2.7) as this is a python2.6 only
dependency. Avoid this trouble by not installing that package version.

I've filed https://github.com/philpep/testinfra/issues/380 upstream to
see if they can fix this.

Change-Id: I5cdfb8467cbc62cf729a9070589f5dcf5c794234