Switch zuul scheduler to new zk cluster
Switch the zuul scheduler to use our new zk cluster. Note this should
happen after the builders have switched and populated image data on the
new database cluster and at the same time as the launchers update.

This gets us off the old single node zk server running on trusty to
newer xenial based zk cluster with HA.

This change does update the config for mergers and executors for
consistency, but I believe that only the scheduler makes use of this
config today.

Change-Id: I132c176af188787cec3a5a562ffdc81f818c26b0
Depends-On: https://review.openstack.org/612441