Configure zk\d to act as a cluster
You have to tell the zk servers what servers belong to the cluster to
have them act as a cluster. Currently all our servers are in mode
standalone as they do not know about the other servers.

Fix this by configuring the three servers as servers 1, 2, and 3 in the
zk config.

Change-Id: I985ac7387744498058f99859eb8445f885b8bbcc