Cleanup Citycloud Sto2
This region does not show up in catalog listings anymore and is causing
inventory generation for ansible to fail. This change removes Sto2 from
the management side of things so that we can get ansible and puppet
running again.

This does not cleanup nodepool which we can do in a followup once
ansible and puppet are running again.

Change-Id: Ifeea238592b897aa4cea47b723513d7f38d6374b