Fix zk cluster members listing
First we convert from a hash to an array to list the zk cluster members.
Do this because the older puppet-zookeeper module we use doesn't support
hashes like newer releases do.

Second we convert from hostnames to IP addresses because the hostnames
were looked up in /etc/hosts to resolve to 127.0.1.1 which caused zk to
listen on 127.0.1.1 and the cluster members could not talk to each
other.

Change-Id: Id590eb535301b27ff0dcb27cf65ba85d61ed19a5