Update clouds.yaml for citycloud with new auth info
citycloud is rolling out per-region keystone. There is a change with an
error in it in the latest openstacksdk, so put the right auth_url into
the files directly while we update it and release it again.

Additoinally, Sto2 and Lon1 each have different domain ids. The domain
names are the same though - and that's good, because logical names are
nicer in config files anyway.

Change-Id: If55d27defc164bd38af2ffd1e7739120389422af