Nodepool.o.o is no longer a thing, remove it
We've only been using nodepool.o.o as a zookeeper server for the past
year or so. Last week we transitioned to a three node zookeeper cluster
and stopped using nodepool.o.o. This server has since been deleted.

This is the last bit of cleanup to remove it from config management.

Change-Id: I9d0363393ed20ee59f40b210ea14fb105a492e20