Update etherpad group membership
Our etherpad servers now have digits so we need them to have digits in
the groups file. Also we have etherpad-dev group_vars that are different
than the etherpad group_vars values so add a group for etherpad-dev to
ensure the correct hiera data is used on the dev server.

Change-Id: I9e34ef7969b8a562192a29d98ead6fc55f71d5f2