Remove mirrors from the webservers group
The webservers group is meant to be mutually exclusive to any other
iptables modifications but was inadvertently expanded to include the
mirror servers by Iea8b2eb2e9c723aca06f75d3d3307893e320cced. Remove
the glob as it should not be there and is overriding the
mirror-specific firewall rules.

Change-Id: Ia1b702e5c9a2d433536edcec4bfaf6f47b1c0291