Remove mirrors from webservers in groups test
We removed the mirror nodes from the webservers group to fix iptables
rule application on the nodes. Unfortunately we didn't update our test
that tries to assert mirrors should be in the webservers group. Update
the test results fixture to remove webservers as a valid group for a
mirror node.

Change-Id: Iba18e54f4df4a36c0247f65642faacca9d195769