Retire the openstack-internships mailing list
This list is basically unused, and its owners/moderators have asked
for it to be shut down. Once this merges (so that Puppet won't
recreate it), the list can be removed with the `rmlist` utility
(this will still leave the archives available but will remove it
from the list index and no longer accept subscriptions/posts).

Change-Id: Ieaf35b1e6875b26b6c20da8a9e9c2eedf200eaf8