bridge.o.o : install ansible 2.7 stable branch
This installs the unreleased stable-branch ansible on bridge.o.o to
incorporate fixes for [1] which is currently stopping the
cloud-launcher from running.

Currently every run it hits citycloud Lon1 and tries to delete it's
router

 TASK [cloud-launcher : Processing router openstackci-router1 for openstackci-citycloud Lon1] ***
 Monday 12 November 2018  04:07:48 +0000 (0:00:00.430)       0:07:45.811 *******
 fatal: [localhost]: FAILED! => {"changed": false, "msg": "Error
 detaching interface from router c7197a8f-096a-4488-a3ae-16fdce0ea580
 ...  cannot be deleted, as it is required by one or more floating
 IPs."}

Although it doesn't succeed, it's probably better that it isn't even
trying...

We will revert when next ansible 2.7 release is out.

[1] https://github.com/ansible/ansible/commit/951572bec1278b00373c9684decd114edf28ed1c

Change-Id: I8f112ba994040c52c7b3c7ee6fd6f5a69fd22919