Stop requiring separate github accounts
This is more of a starting point for discussion. I don't think any of us
actually use a separate github accoutn for github org administration.
For a long time this was probably a mostly non issue because non of us
actually used github for anything. But more recently we have individuals
hacking on ansible and other stuff in github. We should either push
everyone to create a second account for administrative duties, or remove
the rule. This change removes the rule.

Input welcome.

Change-Id: I0c61f192a6b5164af7babde5c99e5ee2b77a652c