Shut down openstack general, dev, ops and sigs mls
The general openstack, openstack-dev, openstack-operators and
openstack-sigs mailing lists have been deprecated since November 19
and are slated to be removed on December 3. Merging this on that
date will ensure any further replies to messages from those lists
are rerouted to the new openstack-discuss mailing list for the
foreseeable future.

The openstack-tc list is included in this batch as it has already
been closed down with a recommendation to send further such
communications to the openstack-discuss ML.

Additionally remove the Puppet mailman resource for the
openstack-sigs ML so it won't be automatically recreated after it
gets deleted (the other lists predate our use of Puppet for this
purpose).

Clean up the corresponding -owner spam rejection aliases since these
addresses will no longer be accepting E-mail anyway.

Change-Id: I9a7fae465c3f6bdcf3ebbadb8926eb4feb8fad79