Add a script to generate the static inventory
We moved from dynamic to static inventory. When creating a new host with
launch-node, a script isn't really needed, the inventory is yaml, the
new host can just be added. However, generating a new inventory by
hitting the APIs of all of our clouds might be useful, so add a utility
script to help in case such a thing is needed.

Change-Id: Iae1be8e9cfe19533005e9f0395d1ef7a6427bc83