Run a local MySQL service on StoryBoard servers
In preparation to move the SB data out of Trove instances to a local
database, make sure the database is created via configuration
management. Make the mysql_backup resource require
storyboard::database instead of storyboard::application to work
around a package resource conflict.

A separate maintenance window will be used to move the data, coupled
with a change of database hostname in private Hiera. Following that,
a followup change will switch the database backups from remote to
local.

Change-Id: I912130ad8a5a139147727e49800e36e2afd815f7