Enable github shared admin account
This change describes the shared github administrator account.

This is inspired by I0c61f192a6b5164af7babde5c99e5ee2b77a652c.  As
described there, this allows for admins to have private accounts in
the organisation, but requires that 2FA be turned on.  If people wish
to keep this as a single account which they do "real" work with
(commits, etc) that is probably OK, but add a note that you'll end up
with a lot of mostly irrelevant stuff in your feeds.

Change-Id: Ic408250571133796b4b4639715fe8d01f91898f2