Add rust-vmm OpenDev ML
A group of interested developers for crosvm and Firecracker are
seeking a neutral location to host a discussion list for
collaboration around secure/special-purpose virtual machine monitors
implemented in the Rust programming language, and asked for the list
name to be rust-vmm. This request came to us via OSF staff who they
approached earlier this week.

Change-Id: I2003e3ae7a352b12d213a553f7ee0cd4f4e5d4f1
Depends-On: https://review.openstack.org/625241