WIP Deploy a kubernetes and a gitea
This is what was used for the POC - but it is NOT in shape to land and
use for real.

percona-xtradb-cluster templates come from the helm chart. We don't
really need to run helm, but it can be used to re-generated the set of
templates. A simple script is included showing how these were generated.

Change-Id: Ie019c2e24c3780cec2468a00987dba4ac34ed570