Only replicate branches and tags to gitea
The refs/changes for repos are a bit expensive size-wise on the
cephfs as the moment (which may be better in the next ceph release)

Aslo, gitea doesn't really know how to handle the refs/changes refs.

For now, only replicate branches and tags.

While we're making settings, bump the thread count up so that we're
replicating a bit more in parallel.

Change-Id: I2ca2834357c6795dc1dfd63c2e39272f51717f76
Depends-On: https://review.openstack.org/629658