Use project-config when syncing gitea repos
We need to create repos in gitea before we do so in Gerrit (because
they will not be replicated correctly if Gerrit pushes to a repo
that doesn't already exist in gitea).  Therefore, we need to add
gitea to the repo creation playbook so that we create new repos
based on the contents of project-config.  This updates the gitea
sync-repos playbook to use that instead of fetching the list from
Gerrit.

It also improves the gitea bootstrap playbook to wait for the load
balancer to come online.

Change-Id: I783a2eed497a830aaf71ad95dea03594774ff6d7