Put opendev website content onto gitea homepage
So that we can host this on opendev.org, we want the content to
be the content from the current website.

This adds a home.tmpl file and puts it into /custom/templates.
Since we're ansibling the content in, we could almost certainly
ansible the md->html conversion - or something similar - but for
now just copied the inner html from the output of that.

Change-Id: Ibe783d743d330f9be2add70e32ad2b6ac36c4111