Add delete-project and importer plugins
delete-project and importer can be combined to achieve renaming
repositories. This is a thing we've wanted to be able to do for
a long time.

The delete-project plugin has an external_plugin_deps.bzl file,
but so does javamelody. We need to combine these into a single
file. Move the ansible into a role so the we can take advantage
of the files directory.

It would probably be nicer to script the combination so we don't
have to maintain this file, but this should work for now.

Change-Id: I30845ca3d8fbb097eb1562e479400245d8cb7d91