Add mirror for shibboleth package for CentOS 7
The keystone team would like to be able to test SAML2.0 federation using
the Shibboleth implementation on more distributions. CentOS/RHEL itself
does not publish Shibboleth packages, but openSUSE happens to host one
of the official mirrors[1]. Since this is mirroring only a few packages
and not an entire distribution, the disk usage should be minimal.

[1] https://shibboleth.net/cgi-bin/mirrorlist.cgi/CentOS_7

Change-Id: I6e7fcd63383256a4cd158c167c7343c1c9c544bf