Put OpenStackID servers in inventory host groups
In order to support different instance names for the OpenStackID
servers, create Ansible inventory host groups for them. Add the
group name for the production server to the Puppet global site
manifest like was already done for the dev server. Also adjust the
production server node name regex and globs to support an enumerated
naming pattern in the openstack.org domain. Finally, fix a benign
missing backslash in the openstackid-dev node name regex for
consistency.

Change-Id: Iaa30f2fa2c196ad84a8ac18b48b7390b1f4905b9