Run base-repos as ubuntu on k8s nodes to bootstrap
k8s-on-openstack uses the baked in ubuntu user and ssh keypairs
to interact with the host. Our other roles assume that we'll be
logging in directly as root.

Run base-repos logging in as ubuntu with become: true set so that
we can overwrite the root ssh key with the one allowing direct
login from bridge.

Change-Id: I98e91e0a9e5f4a44fcad8f22a0f710ce2c4138e0