Remove top level hiera lookups
Puppet4 is a lot more picky about scoping of variables. In our hiera
lookups we rely on the $group var to specify where to find hiera data
files, but this group var is set on a per node basis. This means that
top level hiera lookups produce warnings like:

  Undefined variable 'group';
    (file & line not available)

And lookups will break if they need keys in the group/node specific data
files.

Avoid this problem entirely by removing top level lookups. We only had
one for our elasticsearch_nodes var which we can easily set statically
in site.pp and not rely on hiera for.

Change-Id: Ifa3740f0f31c3fa9a2d8faa355101d3a5671afff