Fix gitea playbooks
First, we need an @ before the extra vars files.  Why?  Because
an @ is needed.

Second, the rook playbook was stringing all 4 commands on to one
exec call which was working poorly.  Instead, make 4 tasks so that
it's slightly better represented in ansible output, each of which
has a (presumably) valid command.

Change-Id: I30efe84d2041237a00da0c0aac02afa92d29c0fb