Disable github rate limit logging
We are currently well under our rate limits for github and logging the
rate limit adds a number of requests to github for each github event
processed. This can be reenabled if we suspect rate limit based
failures, but until then let's reduce the number of requests and
simplify our logging.

Change-Id: I02878f9749737d297680cd324d6dafeefc44a005