Add kdc03.openstack.org
This new Xenial server is being added as a kerberos standby node but
will be used to replace kdc01 as the master once fully configured and
happy as a standby. This replaces the old trusty server.

Note that the server wasn't added to opendev.org as we don't have a
kerberos realm for that domain so that would be a separate activity for
the future.

Change-Id: I4cc5fcd7504c98a7bcd9dc4f2ad57bb5bf8b54bd