Update postgres module to 4.3.0
This module has the puppet 4 (and futureparser) fixes according to the
module changelog [0]. We are trying to switch things to the futureparser
and then puppet 4 so this is a good version to jump to.

Note that the 3.x to 4.x transition is not backward compatible because
the module removed uninstallation features which we don't use. A quick
audit of the interfaces we do consume shows that they haven't changed
but our testing should also check that.

[0] https://github.com/puppetlabs/puppetlabs-postgresql/blob/4.3.0/CHANGELOG.md

Change-Id: Ie2d3afa4f7f2ae3678b17a8b8f7741bf167e14fc