Add --flush-cache to launch.py ansible
I managed to leave off the "--image" flag for a Xenial host, so the
script created a Bionic host by default.  I let that play out, deleted
the host and tried again with the correct image, but what ended up
happening was the fact cache thought this new host was Bionic, and
several ansible roles therefore ran thinking this too, and we ended up
with a bad Xenial/Bionic mashup.

Clear the cache on node launch to avoid this sort of thing again.

I have launched a node with this new option, and it worked.

Change-Id: Ie37f562402bed3846f27fbdd4441b5f4dcec7eb2