Replicate to the standalone gitea severs
We discovered that gitea doesn't like to be run in a shared everything
configuration. Upstream is working to solve those issues but until they
do we'll run 8 independent gitea servers intead of one shared cluster
for "production". The shared cluster remains for testing so we rename
its replication target then add eight new targets for the standalone
servers.

Change-Id: I372627d7fa4240b8d34a9067a2e16dda2b6fdb07