Revert "Revert "Set default branch for repo""
This reverts commit b076aefa0864dc12269a4454a921f4830b2a2d9a.

This may still not work. The last time it got 404s when
running as part of project creation. So possibly it was
failing because there was no git content yet and thus
master was an invalid value. If that's the case, we may
need to add auto_init to true just to get a default repo
and metadata and count on gerrit force-pushing during first
replication.

Change-Id: I66de43a4d00583c06cf92bb7083ce28d4b8d4df9