[WIP] Add unmanaged domains, add acme.opendev.org
This adds the concept of an unmanaged domain; in this case we will
write out the zone file only if it doesn't exist.

acme.opendev.org is added as an unmanaged domain.  It will be managed
by Ansible adding TXT records for ACME authentication.

For testing purposes, we allow passing an extra refspec to the git
checkout.  This can be used to pull in an outsanding review for a CI
run.

Change-Id: I9ef5cfc850c3458c63aff46cfaa0d49a5d194e87
Depends-On: https://review.openstack.org/641154
Depends-On: https://review.openstack.org/641168